+++ test/function/matrix/eye.test.js
@@ -18,7 +18,7 @@ describe('eye', function() {
     assert.deepEqual(math.eye(3,3), matrix([[1,0,0],[0,1,0],[0,0,1]]));
   });
 
-  // TODO: test option math.options.matrix.default
+  // TODO: test option math.options.matrix.defaultType
 
   it('should throw an error with an invalid input', function() {
     assert.throws(function () {math.eye(3,3,2);});