+++ test/function/matrix/zeros.test.js
@@ -51,7 +51,7 @@ describe('zeros', function() {
     assert.deepEqual(zeros([2,3,4]), res);
   });
 
-  // TODO: test option math.options.matrix.default
+  // TODO: test option math.options.matrix.defaultType
 
   it('should create a matrix with zeros with the same size as original matrix', function () {
     var a = matrix([[1, 2, 3], [4, 5, 6]]);